adding border_collision
I added a function to check if the head collided  with the border of the world. For now if the player collide with a border I make sure his position isn't out of the index range of the world and I print a message in the terminal to inform that he touched the border.
I also made the world a rectangle and make the game update faster to check if everything was working properly.