adding the body parts
I added the body part into the game, and I made sure that if I add a body part it will add itself properly.
Before changing the direction of the player, the program now check if it doesn't do a turn-around.
I also added a function to check if the player bumped into itself.